requirements.txt
pytest==6.2.5
pycryptodome==3.10.4
